- Converting this into a gradle project - Adding a docker-compose file to run the loadgen and selenium-chrome together
