+++ .gitignore
@@ -1,0 +1,7 @@
+.idea
+
+# Ignore Gradle project-specific cache directory
+.gradle
+
+# Ignore Gradle build output directory
+build
