+++ app/src/test/java/ai/levo/crapi/loadgen/CrapiLoadgenTest.java
@@ -1,0 +1,12 @@
+/*
+ * This Java source file was generated by the Gradle 'init' task.
+ */
+package ai.levo.crapi.loadgen;
+
+import org.junit.jupiter.api.Test;
+
+class CrapiLoadgenTest {
+    @Test void appHasAGreeting() {
+        CrapiLoadgen classUnderTest = new CrapiLoadgen();
+    }
+}
