+++ build.gradle
@@ -1,0 +1,7 @@
+plugins {
+    id 'java'
+}
+
+subprojects {
+    group = "ai.levo.crapi"
+}
