+++ docker-compose.yml
@@ -1,0 +1,21 @@
+version: "2.1"
+services:
+  selenium-chrome:
+    image: 'selenium/standalone-chrome'
+    ports:
+      - '4444:4444'
+    volumes:
+      - '/dev/shm:/dev/shm'
+    healthcheck:
+      test: curl --fail http://localhost:4444 || exit 1
+      interval: 3s
+      timeout: 10s
+      retries: 5
+  crapi-loadgen:
+    image: 'levoai/crapi-loadgen:latest'
+    depends_on:
+      selenium-chrome:
+        condition: service_healthy
+    environment:
+      CRAPI_URL: '${CRAPI_URL:-http://host.docker.internal:8080}'
+      SELENIUM_HOST: 'selenium-chrome'
