--- pom.xml
@@ -1,63 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <groupId>crAPIlevo</groupId>
-  <artifactId>crAPIDashboard</artifactId>
-  <version>0.0.1-SNAPSHOT</version>
-  <dependencies>
-<dependency>
-    <groupId>org.seleniumhq.selenium</groupId>
-    <artifactId>selenium-java</artifactId>
-    <version>4.1.4</version>
-</dependency>
-  
-  </dependencies>
-  <build>
-    <sourceDirectory>src</sourceDirectory>
-    <plugins>
-      <plugin>
-        <artifactId>maven-compiler-plugin</artifactId>
-        <version>3.8.1</version>
-        <configuration>
-          <release>18</release>
-        </configuration>
-      </plugin>
-     <!-- <plugin>
-      <groupId>org.apache.maven.plugins</groupId>
-      <artifactId>maven-jar-plugin</artifactId>
-      <version>3.1.0</version>
-      <configuration>
-        <archive>
-          <manifest>
-            <addClasspath>true</addClasspath>
-            <classpathPrefix>lib/</classpathPrefix>
-            <mainClass>crAPI.LevocrAPI</mainClass>
-          </manifest>
-        </archive>
-      </configuration>
-      </plugin>-->
-      <plugin>
-      <groupId>org.apache.maven.plugins</groupId>
-    <artifactId>maven-shade-plugin</artifactId>
-    <version>3.0.0</version>
-    <executions>
-       <execution>
-            <phase>package</phase>
-            <goals>
-               <goal>shade</goal>
-            </goals>
-            <configuration>
-                 <transformers>
-                    <transformer 
-implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                        <mainClass>crAPI.LevocrAPI</mainClass>
-                    </transformer>
-                 </transformers>
-             </configuration>
-       </execution>
-    </executions>
-    </plugin>
-    </plugins>
-    
-    <finalName>crapi</finalName>
-  </build>
-</project>
